requirements.txt
fastapi
uvicorn[standard]
requests
pdf2image
pytesseract
pillow
opencv-python-headless
rapidfuzz
numpy
